#487 move reboot button to INFO tab
