+++ frontend/src/components/edge/system/time/host-time.tsx
@@ -9,9 +9,7 @@ export const HostTime = () => {
   const { connUUID = "", hostUUID = "" } = useParams();
   const [data, setData] = useState({} as any);
   const [isFetching, setIsFetching] = useState(false);
-  const [isRebooting, setIsRebooting] = useState(false);
 
-  const hostSystemFactory = new HostSystemFactory();
   const factory = new HostTimeFactory();
   factory.connectionUUID = connUUID;
   factory.hostUUID = hostUUID;
@@ -32,25 +30,6 @@ export const HostTime = () => {
     }
   };
 
-  const rebootHost = async () => {
-    try {
-      setIsRebooting(true);
-      await hostSystemFactory.EdgeHostReboot(connUUID, hostUUID);
-    } finally {
-      setIsRebooting(false);
-    }
-  };
-
-  const warning = () => {
-    Modal.confirm({
-      title: "Confirm Are you sure?",
-      content: "This will reboot the device",
-      onOk() {
-        rebootHost();
-      },
-    });
-  };
-
   return (
     <>
       <Spin spinning={isFetching}>
@@ -66,11 +45,6 @@ export const HostTime = () => {
               <Descriptions.Item label="Time UTC">{data.time_utc}</Descriptions.Item>
             </Descriptions>
             <UpdateTimeSetting />
-            <div className="text-start mt-4">
-              <Button type="primary" onClick={warning} loading={isRebooting} danger>
-                Reboot Host
-              </Button>
-            </div>
           </>
         )}
